# spec/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Habbit Runner API
  version: 1.0.0
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly access-token cookie
    csrfTokenHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF header required for cookie-authenticated mutations
...
  schemas:
...
    CheckinResponseDto:
      type: object
      properties:
        id:
          type: string
        habitId:
          type: string
        date:
          type: string
        done:
          type: boolean
        count:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
        version:
          type: integer
          format: int32
    CheckinUpsertRequestDto:
      type: object
      required:
      - done
      properties:
        done:
          type: boolean
        count:
          type: integer
          format: int32
          exclusiveMinimum: 0
        version:
          type: integer
          format: int32
    CursorPageDtoCheckinResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckinResponseDto"
        nextCursor:
          type: string
    CursorPageDtoHabitResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HabitResponseDto"
        nextCursor:
          type: string
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        errorCode:
          type: string
    HabitColor:
      type: string
      enum:
      - blue
      - green
      - purple
      - orange
      - red
      - cyan
      - "#5E81AC"
      - "#1F2937"
      - "#111827"
    HabitCreateRequestDto:
      type: object
      required:
      - id
      - name
      - color
      - icon
      - frequency
      properties:
        id:
          type: string
          pattern: \S
          maxLength: 100
        name:
          type: string
          pattern: \S
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        color:
          $ref: "#/components/schemas/HabitColor"
        icon:
          type: string
          pattern: \S
          maxLength: 100
        frequency:
          $ref: "#/components/schemas/HabitFrequency"
        customDays:
          type: array
          items:
            type: integer
            format: int32
          maxItems: 7
        schedule:
          $ref: "#/components/schemas/HabitScheduleDto"
        targetStreak:
          type: integer
          format: int32
          exclusiveMinimum: 0
        dailyTarget:
          type: integer
          format: int32
          exclusiveMinimum: 0
        tags:
          type: array
          items:
            type: string
          maxItems: 50
        archived:
          type: boolean
        sortOrder:
          type: integer
          format: int64
        reminderTime:
          type: string
        reminderEnabled:
          type: boolean
        type:
          $ref: "#/components/schemas/HabitType"
        freezeDays:
          type: array
          items:
            type: string
          maxItems: 366
    HabitFrequency:
      type: string
      enum:
      - daily
      - weekdays
      - weekends
      - custom
    HabitResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        color:
          $ref: "#/components/schemas/HabitColor"
        icon:
          type: string
        frequency:
          $ref: "#/components/schemas/HabitFrequency"
        customDays:
          type: array
          items:
            type: integer
            format: int32
        schedule:
          $ref: "#/components/schemas/HabitScheduleDto"
        targetStreak:
          type: integer
          format: int32
        dailyTarget:
          type: integer
          format: int32
        tags:
          type: array
          items:
            type: string
        archived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        version:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int64
        reminderTime:
          type: string
        reminderEnabled:
          type: boolean
        type:
          $ref: "#/components/schemas/HabitType"
        freezeDays:
          type: array
          items:
            type: string
    HabitScheduleDto:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/HabitScheduleType"
        weekdays:
          type: array
          items:
            type: integer
            format: int32
            maximum: 7
            minimum: 1
          maxItems: 7
        timesPerWeek:
          type: integer
          format: int32
          maximum: 7
          minimum: 1
        timesPerMonth:
          type: integer
          format: int32
          maximum: 31
          minimum: 1
        weeksOfMonth:
          type: array
          items:
            $ref: "#/components/schemas/WeekOfMonthValue"
          maxItems: 5
    HabitScheduleType:
      type: string
      enum:
      - daily
      - weekly_days
      - weekly_quota
      - monthly_weeks
      - monthly_quota
    HabitStatusUpdateRequestDto:
      type: object
      required:
      - archived
      properties:
        archived:
          type: boolean
        version:
          type: integer
          format: int32
    HabitType:
      type: string
      enum:
      - positive
      - negative
    HabitUpdateRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        color:
          $ref: "#/components/schemas/HabitColor"
        icon:
          type: string
          maxLength: 100
        frequency:
          $ref: "#/components/schemas/HabitFrequency"
        customDays:
          type: array
          items:
            type: integer
            format: int32
          maxItems: 7
        schedule:
          $ref: "#/components/schemas/HabitScheduleDto"
        targetStreak:
          type: integer
          format: int32
          exclusiveMinimum: 0
        dailyTarget:
          type: integer
          format: int32
          exclusiveMinimum: 0
        tags:
          type: array
          items:
            type: string
        archived:
          type: boolean
        sortOrder:
          type: integer
          format: int64
        reminderTime:
          type: string
        reminderEnabled:
          type: boolean
        type:
          $ref: "#/components/schemas/HabitType"
        freezeDays:
          type: array
          items:
            type: string
          maxItems: 366
        version:
          type: integer
          format: int32
...
    WeekOfMonthValue:
      type: integer
      format: int32
      enum:
      - 1
      - 2
      - 3
      - 4
      - last
...
paths:
...
  /checkins:
    get:
      summary: List checkins
      description: Returns up to the default page of checkins for the authenticated
        user.
      tags:
      - Checkins
      responses:
        "200":
          description: Checkins loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponseDto"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /checkins/habits/{habitId}/dates/{date}:
    put:
      summary: Upsert checkin
      description: Creates or updates a checkin for a concrete habit and date.
      tags:
      - Checkins
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
      - name: habitId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinUpsertRequestDto"
        required: true
      responses:
        "200":
          description: Checkin upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Habit or checkin not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
    delete:
      summary: Delete checkin
      description: Deletes a checkin for a concrete habit and date.
      tags:
      - Checkins
      parameters:
      - name: date
        in: path
        required: true
        schema:
          type: string
      - name: habitId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Checkin deleted
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Habit or checkin not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /checkins/page:
    get:
      summary: Page checkins
      description: Returns checkins using a stable updatedAt/id cursor.
      tags:
      - Checkins
      parameters:
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageDtoCheckinResponseDto"
      security:
      - accessTokenCookie: []
  /habits:
    get:
      summary: List habits
      description: Returns up to the default page of habits for the authenticated
        user.
      tags:
      - Habits
      responses:
        "200":
          description: Habits loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitResponseDto"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      summary: Create habit
      description: Creates or replaces a habit for the authenticated user.
      tags:
      - Habits
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitCreateRequestDto"
        required: true
      responses:
        "201":
          description: Habit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Habit id belongs to another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /habits/page:
    get:
      parameters:
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageDtoHabitResponseDto"
      security:
      - accessTokenCookie: []
      summary: Find Page
      tags:
      - Habit Page Resource
  /habits/{habitId}:
    put:
      summary: Update habit
      description: Updates a habit with a concrete domain-specific request DTO.
      tags:
      - Habits
      parameters:
      - name: habitId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdateRequestDto"
        required: true
      responses:
        "200":
          description: Habit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
    delete:
      summary: Delete habit
      description: Deletes a habit for the authenticated user.
      tags:
      - Habits
      parameters:
      - name: habitId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Habit deleted
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
  /habits/{habitId}/status:
    patch:
      summary: Update habit status
      description: Toggles the habit archive status using a dedicated request DTO.
      tags:
      - Habits
      parameters:
      - name: habitId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitStatusUpdateRequestDto"
        required: true
      responses:
        "200":
          description: Habit status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitResponseDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Habit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - accessTokenCookie: []
...
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value